Move 1 func to `clan_symbols.py`: clan_symbol_sprite
