+++ scripts/screens/LeaderDenScreen.py
@@ -24,9 +24,7 @@ from scripts.screens.Screens import Screens
 from scripts.ui.generate_box import get_box, BoxStyles
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.icon import Icon
-from scripts.utility import (
-    clan_symbol_sprite,
-)
+from scripts.clan_package.clan_symbols import clan_symbol_sprite
 from scripts.ui.theme import get_text_box_theme
 from scripts.events_module.text_adjust import shorten_text_to_fit
 from scripts.clan_package.cotc import get_other_clan, get_other_clan_relation
